changed uint64 to uint32

+++ services/store/client/client.go
@@ -30,7 +30,7 @@ type StoreClientInterface interface {
 	GetUploadStatus(uploadId string) *proto.GetUploadStatusResponse
 	AddRevision(snapName string, sha3_384_encoded string, size uint64, architectures []string, tracksAndChannels []string, unscannedFileName string) *proto.AddRevisionResponse
 	GetObjectCustomMetadata(bucket string, objectKey string) *proto.GetObjectCustomMetadataResponse
-	UpdateUploadStatus(uploadId string, status string, revision uint64, el *cerror.ErrorList) *proto.UpdateUploadStatusResponse
+	UpdateUploadStatus(uploadId string, status string, revision uint32, el *cerror.ErrorList) *proto.UpdateUploadStatusResponse
 }
 
 var _ StoreClientInterface = (*StoreClient)(nil)
@@ -339,7 +339,7 @@ func (c *StoreClient) GetUploadStatus(uploadId string) *proto.GetUploadStatusRes
 // UpdateUploadStatus updates the status of an upload
 // It takes the upload ID, status, revision number, and a list of errors
 // Errors are stored in the database and are retrieved later by Snapcraft to check if any errors occurred during the upload
-func (c *StoreClient) UpdateUploadStatus(uploadId string, status string, revision uint64, el *cerror.ErrorList) *proto.UpdateUploadStatusResponse {
+func (c *StoreClient) UpdateUploadStatus(uploadId string, status string, revision uint32, el *cerror.ErrorList) *proto.UpdateUploadStatusResponse {
 	req := &proto.UpdateUploadStatusRequest{
 		UploadId: uploadId,
 		Status:   status,
